fix ft_uitoa_base.c

+++ ft_uitoa_base.c
@@ -6,19 +6,19 @@
 /*   By: lucocozz <lucocozz@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/22 03:28:55 by lucocozz          #+#    #+#             */
-/*   Updated: 2019/11/22 03:29:25 by lucocozz         ###   ########.fr       */
+/*   Updated: 2019/11/22 03:38:10 by lucocozz         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-char	*ft_uitoa_base(uintmax_t n, char const *base)
+char	*ft_uitoa_base(unsigned int n, char const *base)
 {
-	uintmax_t	len;
-	uintmax_t	div;
-	char		*output;
-	uintmax_t	i;
-	uintmax_t	base_val;
+	unsigned int	len;
+	unsigned int	div;
+	char			*output;
+	unsigned int	i;
+	unsigned int	base_val;
 
 	base_val = ft_strlen(base);
 	div = 1;
